Rename package to remove talend link
