+++ src/main/java/au/org/emii/netcdf/iterator/IndexValuesIterator.java
@@ -1,4 +1,4 @@
-package au.org.emii.talend.dap;
+package au.org.emii.netcdf.iterator;
 
 import java.util.Iterator;
 import java.util.Set;
