+++ src/main/java/au/org/emii/netcdf/iterator/reader/DGridReader.java
@@ -1,4 +1,4 @@
-package au.org.emii.talend.dap.reader;
+package au.org.emii.netcdf.iterator.reader;
 
 import opendap.dap.DGrid;
