+++ src/main/java/au/org/emii/netcdf/iterator/reader/DInt16Reader.java
@@ -1,15 +1,19 @@
-package au.org.emii.talend.dap.reader;
+package au.org.emii.netcdf.iterator.reader;
 
 import java.util.Set;
 
-import opendap.dap.DInt32;
-import au.org.emii.talend.dap.IndexValue;
+import opendap.dap.DInt16;
+import au.org.emii.netcdf.iterator.IndexValue;
 
-public class DInt32Reader extends DReader {
-    private DInt32 variable;
+public class DInt16Reader extends DReader {
+    private DInt16 variable;
     
-    public DInt32Reader(DInt32 variable) {
+    public DInt16Reader(DInt16 variable) {
         this.variable = variable;
+    }
+    
+    public short getInt16(Set<IndexValue> indexValues) {
+        return variable.getValue();
     }
     
     public int getInt32(Set<IndexValue> indexValues) {
@@ -25,8 +29,7 @@ public class DInt32Reader extends DReader {
     }
     
     public String getString(Set<IndexValue> indexValues) {
-        return Integer.toString(variable.getValue());
+        return Short.toString(variable.getValue());
     }
-    
 };
