+++ src/test/java/au/org/emii/netcdf/iterator/utils/FileUtils.java
@@ -1,4 +1,4 @@
-package au.org.emii.talend.dap.utils;
+package au.org.emii.netcdf.iterator.utils;
 
 import java.io.IOException;
